convert snapshotPopup to tsx and add it to the DatasetOverview page
