+++ src/components/dataset/overview/DatasetOverview.jsx
@@ -4,6 +4,7 @@ import { connect } from 'react-redux';
 import { withStyles } from '@mui/styles';
 import { getDatasetById, getDatasetPolicy, getUserDatasetRoles } from 'actions';
 import { Typography } from '@mui/material';
+import SnapshotPopup from 'components/snapshot/SnapshotPopup';
 import DatasetRelationshipsPanel from '../../common/overview/SchemaPanel';
 import { useOnMount } from '../../../libs/utils';
 import { BreadcrumbType, DatasetIncludeOptions } from '../../../constants';
@@ -97,6 +98,7 @@ function DatasetOverview(props) {
           <DatasetOverviewPanel dataset={dataset} />
         </div>
       </div>
+      <SnapshotPopup />
     </div>
   ) : (
     <div />
